修改日志id为uint32

+++ app/app.go
@@ -39,7 +39,7 @@ type appCli struct {
 
 	config core.IConfig
 
-	loggerId uint64
+	loggerId uint32
 	core.ILogger
 
 	component core.IComponent
